requirements.txt
alabaster==0.7.12
appdirs==1.4.3
argh==0.26.2
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
behave==1.2.6
behave-django==1.1.0
behave-webdriver==0.2.2
cached-property==1.5.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coverage==4.0.3
coveralls==1.5.1
defusedxml==0.5.0
deprecation==2.0.6
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-allauth==0.38.0
django-autocomplete-light==3.3.2
django-bootstrap4==0.0.7
django-braces==1.13.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
django-localflavor==2.1
django-multiselectfield==0.1.8
docopt==0.6.2
docutils==0.14
Faker==0.9.1
filelock==3.0.10
flake8==3.6.0
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.8
invoke==1.2.0
isodate==0.6.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.3.0
MarkupSafe==1.1.0
mccabe==0.6.1
mixer==6.1.3
more-itertools==5.0.0
nose==1.3.7
oauthlib==3.0.1
packaging==19.0
parse==1.9.0
parse-type==0.4.2
path.py==11.5.0
pathlib2==2.3.3
pathtools==0.1.2
Paver==1.3.4
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.8.0
prettyconf==1.2.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycep-correios==3.1.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
pyparsing==2.3.1
pytest==4.1.0
pytest-django==3.4.5
python-coveralls==2.9.1
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.8.0
selenium==3.141.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinx-bootstrap-theme==0.6.5
sphinxcontrib-programoutput==0.13
sphinxcontrib-websupport==1.1.0
splinter==0.10.0
static3==0.7.0
text-unidecode==1.2
toml==0.10.0
tox==3.6.1
tox-travis==0.11
typed-ast==1.1.1
urllib3==1.24.1
virtualenv==16.2.0
watchdog==0.9.0
whitenoise==4.1.2
wrapt==1.10.11
zeep==3.2.0
zipp==0.3.3
